Fixing some bugs in the multi-machine script for test mode
This commit has a couple of things. First is it fixes some
bugs in the test mode of the multi-machine scripts. It also
relieves you of running the splitter separately. Even the
splitter.sh script has been removed. The Formatter job is
now run in the reducer which automatically takes care of
the splitting. However, you need to specify an additional
parameter, the number of machines that you want the setup
to run on. The documentation has been updated.

Also, some yahoo specific hard coded paths have been
generalized (Thanks Craig!). The teststream part is removed from the
multi-machine scripts as a mode. This actually makes sense
because when you run in multi-machine mode you are in batch
mode anyways. So why use streaming when just the test mode
can be used.