Failing LDA job on tracker and hadoop failures
Currently, I have seen that during some of these errors
the LDA job doesn't fail and continues to run though one
or few of the trackers have failed and will counted as an
unsuccessful attempt. The restart code will handle it but
failing the job early is better since all of them will be
in sync while the sampling still happens. This needs the
mapred.max.tracker.failures flag to be set for proper
handling by hadoop and mapred.map.max.attempts doesn't
suffice. Fixing that.